refactor(ui): move relevant tab components into settings/account folder
To have them all in one place, adjust references
