+++ app/src/app/[lng]/settings/account/AccountDetailsTabPanel.tsx
@@ -2,8 +2,8 @@ import { FC, useEffect } from "react";
 import { Box, Button, HStack } from "@chakra-ui/react";
 import { SubmitHandler, useForm } from "react-hook-form";
 import { ProfileInputs } from "@/components/GHGI/inventory-pages/settings-page";
-import FormInput from "../../form-input";
-import EmailInput from "../../email-input";
+import FormInput from "../../../../components/form-input";
+import EmailInput from "../../../../components/email-input";
 import { useSetCurrentUserDataMutation } from "@/services/api";
 import { TFunction } from "i18next";
 import { UseSuccessToast } from "@/hooks/Toasts";
