+++ app/src/app/[lng]/settings/account/ManagePasswordTabContent.tsx
